ActiveModeWarden: Add methods for get/request ActiveModeManager's
Also,
a) Includes support for requesting a ClientModeManager for local only
connections. Unlike the other mode managers, this is an opportunistic
use-case, so we should be able to fallback to the primary client mode
manager if STA + STA is not supported. This will be plumbed to
WifiNetworkFactory in a follow up CL.
b) Move WifiNative.initialize() to ActiveModeWarden directly.

Note: Eventually WifiInjector.getClientModeImplHolder() will be
replaced by ActiveModeWarden.getPrimaryClientModeManager.getImpl().
But, that is only possible when we can create disposable ClientModeImpl
instances (which is not the case currently).

Bug: 148935691
Test: atest com.android.server.wifi
Change-Id: I9246746cdb4fe5ffd558aa0df655ba9f51860063