ClientModeImpl: Integrate ConnectingState
We enter the ConnectingState on start of a new connection attempt.
a) We should go to L2ConnectedState on NETWORK_CONNECTION_EVENT received
for the target network.
b) We should go to Disconnected on ASSOCIATION_REJECTION or any other
NETWORK_DISCONNECTION events for the target network.
If any of thes messages are received for another network, ignore it.

Also,
a) Fixed a bunch of unit tests for which expectations were not
correctly set For ex: There are a bunch of tests which send
supplicant state change events without first triggering a connect.
This is not a realistic scenario and breaks with the new changes.
b) Removed testWrongPasswordWithNullNetwork test which is simulating a
situation that cannot happen (getConfiguredNetwork() returns null, but
updateNetworkSelectionStatus() works for the same network?)

Bug: 135503681
Test: Manual tests switching between networks. Ensure from the logs
that we move states as expected:
a) Remain in ConnectingState until we connect to the target network,
ignore any disconnect events for previously connected network.
b) If connection succeeds, move to L2ConnectedState.
b) If connection fails, move to DisconnectedState.

Test: atest com.android.server.wifi
Test: Presubmit ACTS tests.

Change-Id: If2f1ef40edc4590008cd252d55943a0d463cc340