Move setUserConnectChoice from WifiNetworkSelector to WifiConfigManager
WifiConnectivityManager#setUserConnectChoice was simply a pass-through
that delegates to WifiNetworkSelector#setUserConnectChoice.

The only dependency WifiNetworkSelector#setUserConnectChoice has is
WifiConfigManager. Thus the method can be trivially moved to
WifiConfigManager.

Instead of calling WifiConfigManager#saveToStore in both
 #setNetworkConnectChoice & #clearNetworkConnectChoice, only call
 #saveToStore in #setUserConnectChoice since it is the only public API.

This change is necessary because a large part ClientModeImpl#connect
will be moved to WifiConfigManager since it is mode-independent and
can be called when Wifi is off.

Bug: 138403307
Test: atest FrameworksWifiTests
Change-Id: I65eb7af1011c791cfb24a6b6694fc22f264137f9