ClientModeImpl: move mode-independent logic to WifiServiceImpl
For mode-independent logic that is only used by
WifiServiceImpl, it should be moved there.
However, for mode-independent logic that is used
by WifiServiceImpl as well as other classes, we
need to create a new location to put it. This is
because other classes should not call
WifiServiceImpl.

Thus, create ConnectHelper to allow logic to be
shared between WifiServiceImpl & other Wifi
classes.

Also fixed an issue where the
WIFI_CREDENTIAL_CHANGED_ACTION broadcast was not
being sent out upon forgetting a network.

Also only send out the SSID in the
WIFI_CREDENTIAL_CHANGED_ACTION broadcast if
location is enabled.

Bug: 158874479
Bug: 138403307
Test: atest FrameworksWifiTests
Test: connect and forget some networks in Wifi Picker
Change-Id: I41e71663b66218e376240b219743ef4281dcafda