WifiConnectivityManager: Refactor interaction with ClientModeImpl
Refactor WifiConnectivityManager interaction with ClientModeImpl to
handle multiple ClientModeImpl instances.
a) Register for mode change callbacks with ActiveModeWarden to figure
out if wifi is enabled or not.
> 0 primary client mode managers == wifi on
= 0 primary client mode manager == wifi off
b) Listen to screen state changes locally.
c) handleConnectionStateChanged/handleConnectionAttemptEnded now carries
the instance of mode manager associated. If it is not one of the primary
client mode manager instance, we will ignore the call
For ex: WifiConnectivityManager does not need to react to local only
client mode manager state changes for example.

TBD's:
1) Need some more work to handle multiple instances of
primary ClientModeManager.
2) Handle calls from WifiConnectivityManager back into ClientModeImpl.

Bug: 159052883
Test: atest com.android.server.wifi
Test: Presubmit ACTS tests
Change-Id: Ia5f5f8b533ea3eae8e3537c25ef9075ce1980695