ClientModeManager: Remove getImpl()
All the external modules should route their calls via
ClientModeManager. This is needed to create an interface which we can
then use to handle client mode calls when wifi is off.

TODO: Need to create a base class for ClientModeManager and a
DefaultModeManager which returns dummy values when ClientModeManager
instance does not exist (to handle public API calls when wifi is off).

Bug: 159060934
Test: atest com.android.server.wifi
Change-Id: I87b1268080f3223c54e8db8b4470836327715ca9