Populate randomized mac address for passpoint wifi configs
Calculate the randomized mac address for passpoint wifi configs before
returning them in getWifiConfigsForPasspointProfiles. If enhanced mac
randomization is enabled for a config, then populate it with
DEFAULT_MAC_ADDRESS, which displays in Settings as "Not available".

Bug: 160028657
Test: atest PasspointManagerTest, visual verification that randomized
mac address appears in Network Details page for Passpoint networks, and
appears as "Not available" if enhanced mac randomization is enabled.

Change-Id: Icf90940b3c7212f302964db68fc126a012ef36d3