wifi: implement is60GHzBandSupported API in wifi service
Implement WifiManager is60GHzBandSupported for detecting
whether 60GHz band is supported in the system. Note,
this method uses a configuration parameter and does not
check for capable WIFI chip. This is because the API
may need to be called when WIFI is disabled, before
the WIFI HAL is loaded. This can be switched to querying
the HAL in the future, when it is loaded earlier.

Bug: 147469373
Test: atest FrameworksWifiTests
Test: atest FrameworksWifiApiTests
Change-Id: I2fd3383cd6feca05b3824cc1176301222c3ffd7e