ClientModeManager: dynamically instantiate ClientModeImpl
ClientModeImpl will only exist if we are in client mode, not
when we are in scan only mode.

Remove ClientModeImpl.DefaultState which implements a no-op
state, the no-op state is now when ClientModeImpl is null.

Now that ConnectableState is the new base state
in the ClientModeImpl state machine, ensure that it handles
all message types.

Bug: 160335531
Test: atest FrameworksWifiTests
Test: connect to some networks, turn wifi off and on
Change-Id: Iae10e3943bd6bff80e6772865c0a4160ae927655