ClientModeImpl: clean up unused fields
Scanned through all the mutable fields in ClientModeImpl to
check if there is anything that needs to be persisted across
multiple instances of ClientModeImpl (now that it is disposable),
but didn't find anything. Fields that do need to be persisted
are set correctly at initialization, while other fields are
reset upon disconnection and don't need to be persisted.

While here, cleaned up a bunch of unused fields.

Bug: 160335531
Test: atest FrameworksWifiTests
Change-Id: Ie25fdc3a4d2cd8a21885b7ebd282e1a1bdd339a3