WifiDiagnostics: move slow getLogcat outside of synchronized block
Calling getLogcat is very slow (on the order of seconds).
If it is placed inside a synchronized block, it will block all other
methods calls on the WifiDiagnostics object.

Instead, move it outside of the synchronized block, since it is
thread safe and doesn't need to be guarded by a mutex.

Bug: 160446833
Test: atest FrameworksWifiTests
Test: Trigger bug report to force a call to captureBugreportData(),
    then toggle Wifi on to trigger a call to startLogging(). Ensure
    that Wifi isn't stuck toggling on.
Change-Id: Idd10ae0427dcbb31878a4f01fe91ee40ad3b006c