WifiNetworkFactory: Coalesce scan results for matches
Since scans are flaky by nature, use a simple scan cache (with very stale scan
results weeded out). This prevents cases where we don't find a specific
bssid in the latest set of scan resuls (which results in us not setting
the BSSID and hence results in bssid selection in firmware).

Bug: 164441864
Test: atest com.android.server.wifi
Change-Id: I772663f03dcfb5b58fe2f36acec95da19459b111