ClientModeImpl: Ensure we don't end in progress connection event
When connecting to a new network while connected, we should not end the
connection event for the new connection.

We should still end the connection attempt if there is an
NETWORK_DISCONNECTION from the network while the connection is in progress
(most likely for EAP networks, otherwise we should get some L2 failure
like ASSOC-REJECT or AUTH-FAILURE).

Bug: 168242861
Test: atest com.android.server.wifi
Change-Id: I870205e7309db0532327979e0e2003f26a981d51