HalDeviceManager: Add a util for WorkSource operations
Will be used in HalDeviceManager for prioritization of iface requests.

Note: We may need to support > 1 merged worksource for
each iface (For ex: 2 apps requested for aware iface) in the future.
This is not fully plumbed through, but making sure that HalDeviceManager
is ready to handle it if needed in the future.

Bug: 162344695
Test: atest com.android.server.wifi.util
Change-Id: I79c3ac74304030a60ad9dd43cdd96379f34a4353