STA is choosing open mode instead of OWE.
When connecting to AP configured in OWE-TRANSITION mode,
STA is picking open mode instead of OWE.

WiFi framework parse the WFA OWE-TRANSITION
IE and set the OWE-TRANSITION capability in scan capabilities
based on OWE support in wpa-supplicant/driver/firmware. This information
will be available to framework only when supplicant is
running(Connectivity mode). In WiFiNative, OWE feature
support is checked and saved during the first scan operation.
If the device issues a scan in scan only mode, framework fails to query the
OWE feature support as supplicant is not running. This results in
setting the OWE feature to false.

Fixed the issue by setting the OWE feature support when device comes up
in connectivity mode and also when switching from scan mode to connectivity
mode.

Bug: 168816152
Test: Manual - Connect to AP configured in OWE-TRANSITION mode
Change-Id: I75df731385a743a1547d8e8de1c1124e027bc7f5