[WifiTrackerLib] Scanner#start() should do nothing if already started
Scanner#start() should not start a scan loop if it's already started. In
addition, the wifi state should only be read from
WIFI_STATE_CHANGED_ACTION, which will be received after the broadcast
receiver is registered in onStart due to being a sticky broadcast.
Reading the wifi state in onStart() results in a redundant call to
start/stop the scanner.

Bug: 168912608
Test: open wifi picker and observe only one scan issued every 10s from
settings in adb shell dumpsys wifiscanner

Change-Id: Ic545ddd9340776e677f63f3c818bdff75858e551