Mark networks with captive portal status
The goal of this is to eventually enable enhanced randomization by
default on saved open netoworks that do not depend on captive portal.

For new networks, isCaptivePortalNeverDetected should be true until a
captive portal shows up.
For networks that already exist at the upgrade,
isCaptivePortalNeverDetected should be false because there might be a
captive portal that showed up before the upgrade is made.

Bug: 162801581
Test: atest com.android.server.wifi
Test: manual verification of logs.
Change-Id: I3c874078d15228d19d1e611c2aa0409f5d571461