Make app installed cert flags persistent
Store AppInstalledRootCaCert and AppInstalledPrivateKey flags
in the config store to make them persistent. Resolves the issue
of certs that remain on the device after a connection is removed.
The reasone was that these flags were not persistent which caused
the logic to treat the certs as installed by the user.

Bug: 168871784
Test: Manually verify xml file contents
Test: atest XmlUtilTest NetworkListStoreDataTest SupplicantStaNetworkHalTest
Change-Id: Ia174eccd48222837d8565ec4285de17c117743dd