Migrate away from WifiNative#getClientInterfaceName
...to ActiveModeWarden#getPrimaryClientModeManager().

getClientInterfaceName() returns an arbitrary iface,
not necessarily the primary iface.

However, even the primary ClientModeManager may not
be correct for all of these use cases, especially
if the class is concerned with the primary
ClientModeManager changing. This is a TODO for a
later change.

Note: some classes have a circular dependency with
ActiveModeWarden, so instead have to use WifiInjector
to fetch ActiveModeWarden.

Bug: 172086124
Test: atest FrameworksWifiTests
Change-Id: I20ec32c9d67b904b0ed2ca9b668d7c13c088bc5d