ActiveModeWarden: Modify canRequestMoreClientModeManager
Add the requested role in the method to correctly determine if a new STA
is available or not (otherwise we have to wait for the request to
determine if the second STA is available or not even if a particular
use case is fully disabled for the device). This new method will also
be used in a follow up CL to determine sufficiency conditions.

Bug: 169413079
Test: atest com.android.server.wifi
Change-Id: I0937fabe2428ed82f6cb3feb8ebf7ec833afd350