Scan after user disable cellular
Currently, after the user manually selects cell and then disables
cellular data, it may take a while to connect back to wifi if the user
doesn't open or go back to the wifi picker screen.

This shortens the delay the device automatically connects back to wifi
after cellular is disabled.

Bug: 185408942
Test: atest MergedCarrierEntryTest
Change-Id: I1488ba353e514ef86101df22128baccffae9476b