[WifiTrackerLib] Update start info in NetworkDetailsTracker constructor
Info that is populated on onStart() should be updated in the constructor
of the concrete NetworkDetailsTracker classes before Settings can
reference the returned WifiEntry. This prevents Settings from rendering
any UI based on information that may not be ready to reference, such as
security type (resulting in the wrong security used for configuring new
networks) or any data from the network's wificonfiguration (such as
auto-connect on/off, meteredness, mac randomization).

Bug: 187993188
Bug: 186508507
Test: verify WPA3 config is configured when tapping on unsaved WPA3 network
Change-Id: I8c1b0f0c22707897d76d17b55c63df69a6443ec9