Allow different StandardWifiEntry security targeting based on context
StandardWifiEntry security targeting should only target the saved
configurations regardless of scan results in the Saved Networks page and
Network Details page. But for the Wifi Picker and Configure Networks
Page, the targeted security should match the scan results as long as
there are no configs matching those scans. This allows us to configure
new networks when we don't have an available config, or view the saved
config we already have even if we don't have scans for them.

Bug: 187860137
Test: atest WifiTrackerLibTests. Add SAE config when only PSK scans are
available, and verify that the entry still appears in Saved Networks and
Network Details. Verify that the entry is unsaved in the Wifi Picker and
a PSK config will be configured if the network is tapped.

Change-Id: I30ee9462ca69f21b34ee0c4cb1044499a8773c78