Do not give WifiEntry callbacks the WifiEntry lock
Callbacks for WifiEntry should not be given the WifiEntry lock before
the callback method is run. This could result in the lock being held
while code outside of WifiEntry is run, causing deadlock scenarios.

For example:
1) WifiPickerTracker.updateWifiEntries()  acquires WifiPickerTracker
lock to update WifiEntry list.
2) The connected WifiEntry is updated and posts a callback to UI thread
3) UI thread acquires the WifiEntry lock outside of the WifiEntry
methods, runs its own updating logic. If the callback is WifiScanWorker,
it will try to call WifiPickerTracker.getWifiEntries().
4) WifiPickerTracker.getWifiEntries() is blocked on
WifiPickerTracker.updateWifiEntries().
5) WifiPickerTracker.updateWifiEntries() is blocked on the connected
WifiEntry lock, which it needs later in the method. But this is held by
WifiScanWorker, which is blocked on the WifiPickerTracker lock.

Making sure the WifiEntry lock is never held outside of its own methods
will avoid a deadlock scenario.

Bug: 190578653
Test: build
Change-Id: Ifd6922636e94ccd7388e51d3d597155b3382c91c