Create SdkWifiTrackerLib variant of WifiTrackerLib
Create a new library module SdkWifiTrackerLib as a variant of
WifiTrackerLib that is Google3-compatible for SUW. This should be the
exact same as the regular WifiTrackerLib except without any @hide APIs
and with Google3-available dependencies. The intention is for SUW to use
copybara to automatically copy WifiTrackerLib into their project without
needing to modify anything themselves (which was a problem with the
legacy WifiTracker).

Note: WifiNetworkScoreCache is the last remaining @hide API and will be
handled in a separate CL.

Bug: 162368129
Test: m lint-check with sdk_version: "system_current"
Change-Id: I2c947d11de20b14336e7cda61daab89ba6d45a32